+++ x/ibc-hooks/types/msgs.go
@@ -1,0 +1,32 @@
+package types
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+)
+
+// constants.
+const (
+	TypeMsgEmitIBCAck = "emit-ibc-ack"
+)
+
+var _ sdk.Msg = &MsgEmitIBCAck{}
+
+func (m MsgEmitIBCAck) Route() string { return RouterKey }
+func (m MsgEmitIBCAck) Type() string  { return TypeMsgEmitIBCAck }
+func (m MsgEmitIBCAck) ValidateBasic() error {
+	_, err := sdk.AccAddressFromBech32(m.Sender)
+	if err != nil {
+		return sdkerrors.Wrapf(sdkerrors.ErrInvalidAddress, "Invalid sender address (%s)", err)
+	}
+	return nil
+}
+
+func (m MsgEmitIBCAck) GetSignBytes() []byte {
+	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&m))
+}
+
+func (m MsgEmitIBCAck) GetSigners() []sdk.AccAddress {
+	sender, _ := sdk.AccAddressFromBech32(m.Sender)
+	return []sdk.AccAddress{sender}
+}